third_party/WebKit/LayoutTests/editing/execCommand/list-wrapping-image-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(()=> assert_selection(
 '<div contenteditable>|</div>',
  selection => {
  selection.document.execCommand('insertimage');
  selection.document.execCommand('insertorderedlist');
  },
  [
  "<div contenteditable>",
    "<ol>",
      "<li><img>|</li>",
    "</ol>",
  "</div>"),
  'This insertOrderedList on an image in an editable region without selection change.');
</script>
